Extracting some duplicated code
