+++ WordPress/src/main/java/org/wordpress/android/ui/newstats/countries/CountriesCard.kt
@@ -41,15 +41,12 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
+import org.wordpress.android.ui.newstats.components.StatsChangeIndicator
 import org.wordpress.android.ui.newstats.components.StatsGeoChartWebView
+import org.wordpress.android.ui.newstats.components.StatsMapLegend
 import org.wordpress.android.R
 import org.wordpress.android.ui.compose.theme.AppThemeM3
-import androidx.compose.ui.platform.LocalContext
-import androidx.core.content.ContextCompat
-import androidx.compose.ui.graphics.Color as ComposeColor
-import org.wordpress.android.ui.newstats.StatsColors
 import org.wordpress.android.ui.newstats.util.formatStatValue
-import java.util.Locale
 
 private val CardCornerRadius = 10.dp
 private val CardPadding = 16.dp
@@ -197,7 +194,7 @@ private fun LoadedContent(state: CountriesCardUiState.Loaded, onShowAllClick: ()
             Spacer(modifier = Modifier.height(12.dp))
 
             // Legend
-            MapLegend(
+            StatsMapLegend(
                 minViews = state.minViews,
                 maxViews = state.maxViews
             )
@@ -266,46 +263,6 @@ private fun CountryMap(
 }
 
 @Composable
-private fun MapLegend(
-    minViews: Long,
-    maxViews: Long
-) {
-    val context = LocalContext.current
-    // Use the same colors as the map (stats color resources)
-    val colorLow = ComposeColor(ContextCompat.getColor(context, R.color.stats_map_activity_low))
-    val colorHigh = ComposeColor(ContextCompat.getColor(context, R.color.stats_map_activity_high))
-
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            text = formatStatValue(minViews),
-            style = MaterialTheme.typography.labelSmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-        Box(
-            modifier = Modifier
-                .weight(1f)
-                .height(8.dp)
-                .clip(RoundedCornerShape(4.dp))
-                .background(
-                    Brush.horizontalGradient(
-                        colors = listOf(colorLow, colorHigh)
-                    )
-                )
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-        Text(
-            text = formatStatValue(maxViews),
-            style = MaterialTheme.typography.labelSmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-    }
-}
-
-@Composable
 private fun CountryRow(
     country: CountryItem,
     percentage: Float
@@ -371,35 +328,10 @@ private fun CountryRow(
                     fontWeight = FontWeight.SemiBold,
                     color = MaterialTheme.colorScheme.onSurface
                 )
-                ChangeIndicator(change = country.change)
+                StatsChangeIndicator(change = country.change)
             }
         }
     }
-}
-
-@Composable
-private fun ChangeIndicator(change: CountryViewChange) {
-    val (text, color) = when (change) {
-        is CountryViewChange.Positive -> Pair(
-            "+${formatStatValue(change.value)} (${
-                String.format(Locale.getDefault(), "%.1f%%", change.percentage)
-            })",
-            StatsColors.ChangeBadgePositive
-        )
-        is CountryViewChange.Negative -> Pair(
-            "-${formatStatValue(change.value)} (${
-                String.format(Locale.getDefault(), "%.1f%%", change.percentage)
-            })",
-            StatsColors.ChangeBadgeNegative
-        )
-        is CountryViewChange.NoChange -> return
-    }
-
-    Text(
-        text = text,
-        style = MaterialTheme.typography.labelSmall,
-        color = color
-    )
 }
 
 @Composable
