+++ WordPress/src/main/java/org/wordpress/android/ui/newstats/datasource/StatsDataSourceImpl.kt
@@ -220,7 +220,6 @@ class StatsDataSourceImpl @Inject constructor(
                 ReferrersDataResult.Error("Unknown error")
             }
         }
-        return ReferrersDataResult.Error("Referrers API not available")
     }
 
     override suspend fun fetchCountryViews(
